Refactor file structure of tests
