+++ tests/rules/subject-capitalize-first/first-capitalized.input.txt
