+++ tests/rules/subject-capitalize-first/first-capitalized.output.json
