+++ tests/rules/subject-capitalize-first/first-not-capitalized.input.txt
