+++ tests/rules/subject-capitalize-first/first-not-capitalized.output.json
