+++ tests/rules/subject-min-length/less-than-min.input.txt
