+++ tests/rules/subject-min-length/less-than-min.output.json
