+++ tests/rules/subject-min-length/more-than-min.input.txt
