+++ tests/rules/subject-min-length/more-than-min.output.json
