+++ tests/rules/subject-no-end-period/has-end-period.input.txt
