+++ tests/rules/subject-no-end-period/has-end-period.output.json
