+++ tests/rules/subject-no-end-period/no-end-period.output.json
